ML/requirements.txt
fastapi
uvicorn
xgboost
pandas
scikit-learn
psycopg2-binary
python-dotenv
pydantic
joblib
supabase
